Separate classes into their own files
Resolves #6 and #7.
